Split CollisionPluginLoader
To avoid circular dependencies, but enable reuse of the CollisionPluginLoader,
separate the non-ROS part into moveit_core/moveit_collision_detection.so
and the ROS part (reading the plugin name from the parameter server) into
moveit_ros_planning/moveit_collision_plugin_loader.so (as before).
