+++ moveit_core/collision_detection/CMakeLists.txt
@@ -9,6 +9,7 @@ add_library(${MOVEIT_LIB_NAME}
   src/world.cpp
   src/world_diff.cpp
   src/collision_env.cpp
+  src/collision_plugin_loader.cpp
 )
 set_target_properties(${MOVEIT_LIB_NAME} PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
